prevent duplicate store/fetch vehicle
references new VGobjUID variable, VGServerID to detect duplicate get or store requests. this is more reliable, less cumbersome and more efficient than SQF methods of doing this. This should also remove the need to remove gold/gems/coins and returning them to the player when storing the vehicle.

the associated SQF code is outside the scope of this repository, message me on Discord if you want it